Add logging for concurrent http requests

+++ src/http/ConnectionRateLimiter.ts
@@ -1,4 +1,7 @@
 import {Request, Response} from "express";
+import {getLogger} from "@atomiqlabs/lp-lib/dist/utils/Utils";
+
+const logger = getLogger("ConcurrentRequestsLimitter: ");
 
 export function createConnectionRateLimiter(maxConnections?: number): (req: Request, res: Response, next: () => void) => void {
     maxConnections ??= 5;
@@ -7,22 +10,32 @@ export function createConnectionRateLimiter(maxConnections?: number): (req: Requ
 
     return (req: Request, res: Response, next: () => void) => {
         const ip = req.ip;
-        let connectionCount = (ipStore.get(ip) ?? 0) + 1;
-        ipStore.set(ip,  connectionCount);
+        let connectionCount = ipStore.get(ip) ?? 0;
 
-        res.on('finish', () => {
+        if (connectionCount >= maxConnections) {
+            res.status(503).send('Too many open requests');
+            return;
+        }
+
+        ipStore.set(ip,  connectionCount+1);
+        logger.debug("OPEN "+ip+" -> "+req.path+", concurrent connections: "+(connectionCount+1));
+
+        let cleanedUp = false;
+        const cleanup = () => {
+            if(cleanedUp) return;
+            cleanedUp = true;
             let connectionCount = (ipStore.get(ip) ?? 1) - 1;
+            logger.debug("CLOSE "+ip+" -> "+req.path+", concurrent connections: "+connectionCount);
             if(connectionCount===0) {
                 ipStore.delete(ip);
             } else {
                 ipStore.set(ip, connectionCount);
             }
-        });
+        };
 
-        if (connectionCount > maxConnections) {
-            res.status(503).send('Too many open connections');
-            return;
-        }
+        res.on('finish', cleanup);
+        res.on('error', cleanup);
+        res.on('close', cleanup);
 
         next();
     };
